+++ test/files/sqlite/schema.sql
@@ -12,6 +12,7 @@ create table videos (
 	user_id int not null,
 	sponsor_id int unique,
 
+    unique (user_id, sponsor_id),
 	foreign key (user_id) references users (id),
 	foreign key (sponsor_id) references sponsors (id)
 );
@@ -98,8 +99,7 @@ create table type_monsters (
 	time_nine   timestamp not null default current_timestamp,
 	time_eleven date null,
 	time_twelve date not null,
-	time_fifteen date null default '19990108',
-	time_sixteen date not null default '1999-01-08',
+	time_fifteen date not null default '1999-01-08',
 
 	json_null  json null,
 	json_nnull json not null,
@@ -193,6 +193,38 @@ CREATE TABLE has_generated_columns (
    e TEXT GENERATED ALWAYS AS (substr(c,b,b+1)) STORED
 );
 
+CREATE TABLE test_index_expressions (
+    col1 int,
+    col2 int,
+    col3 int
+);
+CREATE INDEX idx1 ON test_index_expressions ((col1 + col2));
+CREATE INDEX idx2 ON test_index_expressions ((col1 + col2), col3);
+CREATE INDEX idx3 ON test_index_expressions (col1, (col2 + col3));
+CREATE INDEX idx4 ON test_index_expressions (col3);
+CREATE INDEX idx5 ON test_index_expressions (col1 DESC, col2 DESC);
+CREATE INDEX idx6 ON test_index_expressions (POW(col3, 2));
+
+CREATE TABLE foo_bar (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    secret_col TEXT NOT NULL
+);
+CREATE TABLE foo_baz (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    secret_col TEXT NOT NULL
+);
+CREATE TABLE foo_qux (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    secret_col TEXT NOT NULL
+);
+CREATE TABLE bar_baz (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    secret_col TEXT NOT NULL
+);
+CREATE TABLE bar_qux (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    secret_col TEXT NOT NULL
+);
 
 
 -- For the attached database
@@ -250,39 +282,6 @@ create table one.as_generated_columns (
    c TEXT,
    d INT GENERATED ALWAYS AS (a*abs(b)) VIRTUAL,
    e TEXT GENERATED ALWAYS AS (substr(c,b,b+1)) STORED
-);
-
-CREATE TABLE test_index_expressions (
-    col1 int,
-    col2 int,
-    col3 int
-);
-CREATE INDEX idx1 ON test_index_expressions ((col1 + col2));
-CREATE INDEX idx2 ON test_index_expressions ((col1 + col2), col3);
-CREATE INDEX idx3 ON test_index_expressions (col1, (col2 + col3));
-CREATE INDEX idx4 ON test_index_expressions (col3);
-CREATE INDEX idx5 ON test_index_expressions (col1 DESC, col2 DESC);
-CREATE INDEX idx6 ON test_index_expressions (POW(col3, 2));
-
-CREATE TABLE foo_bar (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    secret_col TEXT NOT NULL
-);
-CREATE TABLE foo_baz (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    secret_col TEXT NOT NULL
-);
-CREATE TABLE foo_qux (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    secret_col TEXT NOT NULL
-);
-CREATE TABLE bar_baz (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    secret_col TEXT NOT NULL
-);
-CREATE TABLE bar_qux (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    secret_col TEXT NOT NULL
 );
 
 CREATE TABLE one.foo_bar (